pyproject.toml
[tool.poetry]
name = "gnn_gpu"
version = "0.1.0"
description = "project for testing GNNs and GPUs"
authors = ["uhTVik <uhatov007@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
vk = "^2.0.2"
vk-api = "^11.9.5"
pandas = "^1.3.4"
googletrans = "3.1.0a0"
facebook-sdk = "^3.1.0"
networkx = "^2.6.3"
matplotlib = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
